+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 
 ### Removed
 
+- Suppression du test sur [mbstring.overload](https://www.php.net/manual/en/mbstring.configuration.php#ini.mbstring.func-overload) à l'install
 - Suppression de l’usage de la classe `no_image_filtrer` (utiliser la classe `filtre_inactif` qui l’a remplacé)
 - Fichier `plugins-dist.json` (on utilise `composer.json` maintenant)
 - Boucle `POUR` (utiliser une boucle DATA tel que `<BOUCLE_x(DATA){source tableau, ...}>`)